Refactor out JodChild
