+++ crates/flycheck/Cargo.toml
@@ -17,3 +17,4 @@ serde_json = "1.0.48"
 jod-thread = "0.1.1"
 
 toolchain = { path = "../toolchain", version = "0.0.0" }
+stdx = { path = "../stdx", version = "0.0.0"  }
